move dataflow::Client to dataflow_types
