+++ src/dataflow/src/lib.rs
@@ -24,7 +24,4 @@ mod sink;
 pub mod logging;
 pub mod source;
 
-pub use server::{
-    serve, Client, Command, Config, LocalClient, Response, Server, TimestampBindingFeedback,
-    WorkerFeedback,
-};
+pub use server::{serve, Config, Server};
