+++ src/dataflowd/Cargo.toml
@@ -8,6 +8,7 @@ publish = false
 [dependencies]
 anyhow = "1.0.45"
 async-trait = "0.1.51"
+dataflow-types = { path = "../dataflow-types" }
 dataflow = { path = "../dataflow" }
 log = "0.4.13"
 ore = { path = "../ore" }
